Add React Router and Supabase auth functionality
In App.js, React Router functionality is added
through redirecting the users to respective
URLs (/login and /home) respectively. Loading
anywhere in the site will redirect to either
of these two pages depending on authentication
status with Supabase.

Basic Supabase functionality is added, with
Google account being a work in progress. When
a user enters the app, they also have an option
to log out of the app if needed.