Fix Google integration and add loading screen
Google integration is fixed by adding another
environment variable which represents the
URL of the app. For local machines, this will
be http://localhost:3000/dashboard/. For
new deployments, use the URL
https://todo-study-deployment-url/dashboard.

As automatic redirects lead to the /dashboard
page, a simple Material UI loading screen is
added to allow the page to load.