Add change account page, refactor AvatarView
Add change account page using the same AccountSettings component
as the account creation page. Allows for change in username if
username has not been taken.

AvatarView is refactored such that it does not require the use of
username, firstName, and lastName props all the time. The instance
where AvatarView is used for account display is also accounted for
in the AccountSettings component, where the src prop sets the avatar
to the new one upon the upload of a new avatar.