Remove error alert and refactor fetchUserInfo
As fetching the user info is reused between different components,
exporting the fetchUserInfo in its own page helps to reduce
boilerplate code.

Removed the error alert which was the main cause of the bug.
As the page will automatically refresh/redirect once there is an
error in retrieving data, there is really no need for a proper
alert dialog, and this is placed in the console log instead.

Added a functionality to remove an already placed avatar in
the account settings page.