Add UI to timer and primary timer functionality
The timer app works in repeated work-rest cycles to accommodate
for study techniques like Pomodoro. To act like a normal timer
that counts down, you can set the timer to:

          { work: <time>, rest: 0, cycles: 1 }

This timer functionality also includes the TimeUpDialog, which
appears once the timer has finished.

Notable issues with this timer is that it does not sustain beyond
page visits and a refresh will reset the entire timer progress.
This will be worked on in the future, likely using localStorage as
people aren't going to transfer the timer between devices. However,
this functionality could be implemented in the future as an
additional convenience feature.

A bug with this timer occurs when testing the app through React. A
React state refresh could cause the setInterval() to go haywire and
potentially never stop. This can be fixed by ALWAYS refreshing the page
before testing the timer (although the localStorage option may render
this more difficult).