Fix bug in account settings avatar feature
In the account settings page after the account is created, a bug occurs
when you remove the avatar then set a new avatar again, as it causes
the avatar to display the plain avatar. This is fixed here.

TODO: add new functionality to retain avatar to old one (as uploaded
into server - an easy way could be to click on the back to dashboard
button and go back to the account settings page so this isn't an
essential feature)