Reorganise files and fix bugs
Organised the files for each component into folders and renamed
them as fit. Prepared files and components for the implementation of
the note-taking app.

Tried to fix the issue within the timer where it sends too much data
too frequently to the Supabase server which resulted in a data overrun
and fixed most warnings generated by ESLint.

Added a message to the timer template selection dialog informing users
that it will take a few seconds for the new timer template to be loaded.