Fix alert when accessing timer for first time
When accessing the timer app for the first time, since the timer
settings have to be initialised for first-time users, there is a need
to create the entry representing a user's timer settings. However,
this may lead to instances where an alert pop up showing an error
when inserting the user's default details onto the database.

Upon testing, it turns out the user's timer settings is created and
inserted into the database just fine, so the unnecessary error alert
has been removed. If there are any error, it is likely the page would
rerender until the insertion is successful.

Added documentation for some of the React components. Documentation
will eventually be added for the remaining React components.