Change accordion to dialog
Change the accordion to dialog as the accordion for the filters takes
up nearly the entire height of the dialog, which isn't good especially
for mobile devices. Used a separate dialog instead. Also put the filter
dialog into a separate file.

TODO: ensure the fields don't clear when the dialog opens