Add validation and minor styling changes
Added validation to the InputTimerField component such that it does not
allow any decimal values or values that are out of reasonable range
(e.g. 61 minutes is no longer allowed as an input as this can also be
represented as 1 hour 1 minute).

Made minor styling changes, especially to the buttons and the search
timer templates functionality.