Fix layout and scroll issues
Height of the MainEditor component (on the right side) is fixed to match
the one in the NoteNavigation (on the left side) in the Notes app.

Due to the pop-ups being part of the editor package and being impossible
to customise using JavaScript (even the package creator are using band-
aid ways to deal with this problem), there is a need to account for
pop-ups cutting out of the screen. Fixed the scroll-bar issue where it
is impossible to scroll right in this scenario due to the floating
action button in the bottom right blocking the way. With the new
combined React hook, position of the floating action button can be set
back to 'absolute', completely resolving this issue.