Fix export for mobile (iOS) devices
Modify the exporting algorithm such that a new window/popup opens and
prints from there instead, allowing for export in iOS devices. Prior
to this, the export as PDF option results in the print preview showing
the entire Notes app instead of the individual Note for iOS devices.

TODO: Check if export as PDF works for Android devices